refactor: replace RenderVideoModel with VideoRenderData across platform files
