+++ lib/core/platform/platform_interface.dart
@@ -7,8 +7,8 @@ import '/core/models/thumbnail/key_frames_configs_model.dart';
 import '/core/models/thumbnail/thumbnail_configs_model.dart';
 import '/core/models/video/editor_video_model.dart';
 import '/core/models/video/progress_model.dart';
-import '/core/models/video/render_video_model.dart';
 import '/core/models/video/video_metadata_model.dart';
+import '../models/video/video_render_data_model.dart';
 import 'native_method_channel.dart';
 
 /// Abstract platform interface for the Pro Video Editor plugin.
