+++ lib/pro_video_editor.dart
@@ -3,7 +3,7 @@
 export 'core/models/video/progress_model.dart';
 export 'core/models/video/editor_video_model.dart';
 export 'core/models/video/export_transform_model.dart';
-export 'core/models/video/render_video_model.dart';
+export 'core/models/video/video_render_data_model.dart';
 export 'core/models/video/video_segment_model.dart';
 export 'core/models/video/video_metadata_model.dart';
 export 'core/models/video/video_quality_preset.dart';
